refactor: move game:command ipc sub to context; page uses pubsub service
